backend/requirements.txt
# Core Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
pydantic-settings>=2.6.0

# Database
sqlalchemy[asyncio]>=2.0.36
alembic>=1.14.0
asyncpg>=0.30.0

# Redis
redis[hiredis]>=5.2.0

# Polymarket
py-clob-client>=0.18.0

# Web3 & Cryptography
eth-account>=0.13.0
eth-utils>=5.1.0

# Security & Auth
PyJWT>=2.10.0
cryptography>=44.0.0
python-multipart>=0.0.18

# Async HTTP
httpx>=0.28.0

# Background Tasks
APScheduler>=3.10.4

# Data
python-dateutil>=2.9.0

# Monitoring & Logging
python-json-logger>=3.2.0

# Development
ruff>=0.8.0
mypy>=1.13.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-cov>=6.0.0
